Browser open implementation didn't work out, but retain preparatory refactoring
